--- src/resources/status/controller.py
@@ -1,12 +1,0 @@
-from flask_restful import Resource
-
-
-class StatusController(Resource):
-    url = "/status"
-
-    def get(self):
-        return {"msg": "Status OK"}
-
-
-def add_status_resource_table(api):
-    api.add_resource(StatusController, StatusController.url)
